requirements.txt
flask==2.3.3
flask-login==0.6.2
werkzeug
python-dotenv>=1.0.0
requests==2.31.0
discord.py>=2.3.2
SQLAlchemy>=2.0.0
bcrypt==4.0.1
PyQt6>=6.5.0
pyinstaller>=5.13.0